% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converter.R
\name{convert_csv}
\alias{convert_csv}
\title{Convert XML to CSV}
\usage{
convert_csv(
  year,
  form_type,
  recipient = FALSE,
  schedule = "",
  output_file,
  nrows = NULL
)
}
\arguments{
\item{year}{The year of the data.}

\item{form_type}{The IRS form type to process.}

\item{recipient}{Boolean to extract recipient data.}

\item{schedule}{The schedule to extract, default is the index data.}

\item{output_file}{The path where the CSV file will be saved.}

\item{nrows}{The number of rows to read from the index file. Default is unlimited (NULL).}
}
\description{
This function extracts IRS form data from XML files and saves it as a CSV file.
}
\examples{
# Example CSV file paths (ensure these files exist in your package's inst/variables/ directory)
all_variables_file <- system.file("variables/all_variables.csv", package = "ConverterIRS990")
recipient_variables_file <- system.file("variables/recipient_variables.csv", package = "ConverterIRS990")
schedule_c_variables_file <- system.file("variables/schedule_c_variables.csv", package = "ConverterIRS990")

# If the files exist, run the example
if (file.exists(all_variables_file) && file.exists(recipient_variables_file) && file.exists(schedule_c_variables_file)) {
  convert_csv(2024, "990", FALSE, "", "output_file.csv", nrows = 100)
}
}
