% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converter.R
\name{convert_rda}
\alias{convert_rda}
\title{Convert XML to RDA}
\usage{
convert_rda(
  year,
  form_type,
  recipient = FALSE,
  schedule = "",
  output_file,
  nrows = NULL
)
}
\arguments{
\item{year}{The year of the data.}

\item{form_type}{The IRS form type to process.}

\item{recipient}{Boolean to extract recipient data.}

\item{schedule}{The schedule to extract, default is the index data.}

\item{output_file}{The path where the RDA file will be saved.}

\item{nrows}{The number of rows to read from the index file. Default is unlimited (NULL).}
}
\description{
This function extracts IRS form data from XML files and saves it as an RDA file.
}
\examples{
convert_rda(2024, "990", FALSE, "", "output_file.rda", nrows = 100)
}
