% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloader.R
\name{download_and_extract_zip}
\alias{download_and_extract_zip}
\title{Download and Extract ZIP}
\usage{
download_and_extract_zip(xml_files_path_prefix, xml_batch_id, year)
}
\arguments{
\item{xml_files_path_prefix}{The local path prefix for XML files.}

\item{xml_batch_id}{The batch ID for the XML files.}

\item{year}{The year of the data.}
}
\description{
This function downloads and extracts ZIP files from IRS servers.
}
\keyword{internal}
