% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloader.R
\name{download_and_extract_zip_legacy}
\alias{download_and_extract_zip_legacy}
\title{Download and Extract ZIP for Legacy Data}
\usage{
download_and_extract_zip_legacy(xml_files_path, year)
}
\arguments{
\item{xml_files_path}{The local path for XML files.}

\item{year}{The year of the data.}
}
\description{
This function downloads and extracts ZIP files for legacy data (pre-2024).
}
\keyword{internal}
