% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloader.R
\name{download_index_csv}
\alias{download_index_csv}
\title{Download Index CSV}
\usage{
download_index_csv(year)
}
\arguments{
\item{year}{The year of the data to download.}
}
\value{
The local path to the downloaded CSV file.
}
\description{
This function downloads the index CSV file for a specific year.
}
\keyword{internal}
