% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{extract_index_data}
\alias{extract_index_data}
\title{Extract Index Data}
\usage{
extract_index_data(year, form_type, nrows = NULL)
}
\arguments{
\item{year}{The year of the data.}

\item{form_type}{The form type to filter on.}

\item{nrows}{The number of rows to read from the index file. Default is unlimited (NULL).}
}
\value{
A data frame containing the extracted data.
}
\description{
This function extracts data from the index file for a specific year and form type. # nolint: line_length_linter.
}
\keyword{internal}
