% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{extract_schedule_c_data}
\alias{extract_schedule_c_data}
\title{Extract Schedule C Data}
\usage{
extract_schedule_c_data(year, nrows = NULL)
}
\arguments{
\item{year}{The year of the data.}

\item{nrows}{The number of rows to read from the index file. Default is unlimited (NULL).}
}
\value{
A data frame containing the extracted Schedule C data.
}
\description{
This function extracts Schedule C data from XML files for a specific year.
}
\keyword{internal}
