% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_variables_and_attr_from_xml}
\alias{extract_variables_and_attr_from_xml}
\title{Extract Variables and Attributes from XML}
\usage{
extract_variables_and_attr_from_xml(xml_file, variables)
}
\arguments{
\item{xml_file}{The path to the XML file.}

\item{variables}{A vector of variables to extract.}
}
\value{
A list of extracted variables and attributes.
}
\description{
This function extracts specific variables and their attributes from an XML file.
}
\keyword{internal}
