% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_variables_from_csv}
\alias{read_variables_from_csv}
\title{Read Variables from CSV}
\usage{
read_variables_from_csv(file_path)

read_variables_from_csv(file_path)
}
\arguments{
\item{file_path}{Path to the CSV file.}
}
\value{
A character vector of variables.

A character vector of variables.
}
\description{
This function reads a list of variables from a CSV file.

This function reads a list of variables from a CSV file.
}
\keyword{internal}
